Consolidate conn fixture definitions
